Move pregel independent graph stuff into separate directory
