+++ arangod/Pregel/CMakeLists.txt
@@ -7,7 +7,6 @@ add_library(arango_pregel STATIC
   OutgoingCache.cpp
   PregelFeature.cpp
   PregelMetrics.cpp
-  PregelOptions.cpp
   Utils.cpp)
 
 target_link_libraries(arango_pregel
