+++ arangod/Pregel/Collections/Graph/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(arango_pregel PRIVATE
+  Collections.cpp
+  Properties.cpp
+  Source.cpp)
