+++ arangod/Pregel/Conductor/GraphSource.cpp
@@ -1,73 +1,13 @@
 #include "GraphSource.h"
 
 #include "Cluster/ServerState.h"
-#include "Pregel/Collections/CollectionFactory.h"
-#include "Pregel/PregelOptions.h"
+#include "Pregel/Collections/Graph/Collections.h"
 #include "VocBase/vocbase.h"
 #include "fmt/ranges.h"
 
-using namespace arangodb;
-using namespace arangodb::pregel;
 using namespace arangodb::pregel::conductor;
 
-namespace {
-template<class... Ts>
-struct overloaded : Ts... {
-  using Ts::operator()...;
-};
-template<class... Ts>
-overloaded(Ts...) -> overloaded<Ts...>;
-}  // namespace
-
-auto GraphCollections::convertToShards(
-    EdgeCollectionRestrictions const& restrictions) const
-    -> std::unordered_map<VertexShardID, std::vector<EdgeShardID>> {
-  auto restrictionsPerShard =
-      std::unordered_map<VertexShardID, std::vector<EdgeShardID>>{};
-  for (auto const& [vertexCollectionId, edgeCollectionIds] :
-       restrictions.items) {
-    auto vertexCollection = vertexCollections.find(vertexCollectionId);
-    if (vertexCollection == std::nullopt) {
-      continue;
-    }
-    for (auto const& vertexShard : vertexCollection.value()->shards()) {
-      for (auto const& edgeCollectionId : edgeCollectionIds) {
-        auto edgeCollection = edgeCollections.find(edgeCollectionId);
-        if (edgeCollection == std::nullopt) {
-          continue;
-        }
-        for (auto const& edgeShard : edgeCollection.value()->shards()) {
-          restrictionsPerShard[vertexShard].emplace_back(edgeShard);
-        }
-      }
-    }
-  }
-  return restrictionsPerShard;
-}
-
-auto GraphCollections::from(GraphCollectionNames const& names,
-                            TRI_vocbase_t& vocbase)
-    -> ResultT<GraphCollections> {
-  auto collectionFactory = collections::CollectionFactory{vocbase};
-  auto vertexCollections = collectionFactory.create(names.vertexCollections);
-  if (vertexCollections.fail()) {
-    return vertexCollections.result();
-  }
-  auto edgeCollections = collectionFactory.create(names.edgeCollections);
-  if (edgeCollections.fail()) {
-    return edgeCollections.result();
-  }
-  return GraphCollections{.vertexCollections = vertexCollections.get(),
-                          .edgeCollections = edgeCollections.get()};
-}
-
-auto GraphCollections::all() const -> collections::Collections {
-  auto allCollections = vertexCollections;
-  allCollections.insert(edgeCollections);
-  return allCollections;
-}
-
-auto GraphSourceSettings::isShardingCorrect(
+auto GraphSettings::isShardingCorrect(
     std::shared_ptr<collections::Collection> const& collection) -> Result {
   if (!collection->isSmart()) {
     std::vector<std::string> eKeys = collection->shardKeys();
@@ -86,14 +26,14 @@ auto GraphSourceSettings::isShardingCorrect(
   return Result{};
 }
 
-auto GraphSourceSettings::getSource(TRI_vocbase_t& vocbase)
+auto GraphSettings::getSource(TRI_vocbase_t& vocbase)
     -> ResultT<PregelGraphSource> {
   auto graphCollectionNames = graphDataSource.collectionNames(vocbase);
   if (graphCollectionNames.fail()) {
     return graphCollectionNames.result();
   }
-  auto graphCollections =
-      GraphCollections::from(graphCollectionNames.get(), vocbase);
+  auto graphCollections = collections::graph::GraphCollections::from(
+      graphCollectionNames.get(), vocbase);
   if (graphCollections.fail()) {
     return graphCollections.result();
   }
